requirements.txt
asgiref==3.8.1
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.3.2
cloudinary==1.41.0
constantly==23.10.4
cryptography==43.0.1
daphne==4.1.2
Django==5.1.1
django-cloudinary-storage==0.3.0
django-cors-headers==4.4.0
django-debug-toolbar==4.4.6
django-filter==24.3
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
ffmpeg-python==0.2.0
future==1.0.0
google-api-core==2.19.2
google-api-python-client==2.144.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.65.0
httplib2==0.22.0
hyperlink==21.0.0
idna==3.8
incremental==24.7.2
msgpack==1.1.0
packaging==24.1
pillow==10.4.0
proto-plus==1.24.0
protobuf==5.28.0
psycopg2==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.9.0
pyOpenSSL==24.2.1
pyparsing==3.1.4
python-decouple==3.8
razorpay==1.4.2
redis==5.0.8
requests==2.32.3
rsa==4.9
service-identity==24.1.0
setuptools==74.1.2
six==1.16.0
sqlparse==0.5.1
Twisted==24.7.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
zope.interface==7.0.3
